+++ packages/@cdk-cosmos/core/src/cicd-solar-system/index.ts
@@ -1,0 +1,12 @@
+export {
+  ICiCdSolarSystemCore,
+  CiCdSolarSystemCoreStack,
+  CiCdEcsSolarSystemCoreStack,
+  CiCdSolarSystemCoreStackProps,
+} from './cicd-solar-system-core-stack';
+export {
+  ICiCdSolarSystemExtension,
+  CiCdSolarSystemExtensionStack,
+  CiCdEcsSolarSystemExtensionStack,
+  CiCdSolarSystemExtensionStackProps,
+} from './cicd-solar-system-extension-stack';