+++ packages/@cdk-cosmos/core/src/cosmos/index.ts
@@ -1,0 +1,3 @@
+export { ICosmosCore, CosmosCoreStack, CosmosCoreStackProps } from './cosmos-core-stack';
+export { CosmosCoreImport, CosmosCoreImportProps } from './cosmos-core-import';
+export { ICosmosExtension, CosmosExtensionStack, CosmosExtensionStackProps } from './cosmos-extension-stack';