+++ packages/@cdk-cosmos/core/src/helpers/utils.ts
@@ -1,3 +1,5 @@
+import * as fs from 'fs';
+import * as path from 'path';
 import { Stack, IConstruct } from '@aws-cdk/core';
 
 export const isCrossAccount = (x: IConstruct, y: IConstruct, includeRegion?: boolean): boolean => {
@@ -6,4 +8,9 @@ export const isCrossAccount = (x: IConstruct, y: IConstruct, includeRegion?: boo
   const diffAccount = stackX.account !== stackY.account;
   const diffRegion = stackX.region !== stackY.region;
   return includeRegion ? diffAccount && diffRegion : diffAccount;
+};
+
+export const getPackageVersion: () => string = () => {
+  const file = fs.readFileSync(path.resolve(__dirname, '../../package.json')).toString();
+  return JSON.parse(file).version as string;
 };