+++ packages/@cdk-cosmos/core/test/solar-system.test.ts
@@ -11,6 +11,7 @@ import {
   SolarSystemCoreStack,
   SolarSystemExtensionStack,
 } from '../src';
+import { SolarSystemCoreImport } from '../src/solar-system';
 
 const app = new App();
 const env = { account: 'account', region: 'region' };
@@ -90,7 +91,7 @@ describe('Solar-System', () => {
     }).toThrowError('NetworkBuilder not found, please define cidr range here (SolarSystem: Test) or Galaxy or Cosmos.');
   });
 
-  test('should have cird range', () => {
+  test('should have cidr range', () => {
     let app = new App();
     let cosmos = new CosmosCoreStack(app, 'Test', { tld: 'com', cidr: '10.0.0.0/22' });
     let galaxy = new GalaxyCoreStack(cosmos, 'Test');
@@ -183,8 +184,9 @@ describe('SolarSystem Extension', () => {
     const app = new App();
     const cosmos = new CosmosExtensionStack(app, 'Test', { env });
     const galaxy = new GalaxyExtensionStack(cosmos, 'Test');
-    const sys = new SolarSystemExtensionStack(galaxy, 'Test', {});
-    const sys2 = new SolarSystemExtensionStack(galaxy, 'Test2', { portalProps: { id: 'Test' } });
+    const portal = new SolarSystemCoreImport(galaxy.portal, 'Test');
+    const sys = new SolarSystemExtensionStack(galaxy, 'Test', { portal });
+    const sys2 = new SolarSystemExtensionStack(galaxy, 'Test2', { portal });
     new ARecord(sys, 'Test', {
       zone: sys.portal.zone,
       target: RecordTarget.fromIpAddresses('1.1.1.1'),