+++ packages/@cosmos-building-blocks/pipeline/src/index.ts
@@ -1,1 +1,4 @@
-export { AppNodePipeline, AppNodePipelineProps } from './app-node-pipeline';
+export { BuildSpecObject, BuildSpecBuilder } from './build-spec';
+export { StandardPipeline, StandardPipelineProps } from './standard-pipeline';
+export { NodePipeline, NodePipelineProps } from './node-pipeline';
+export { DockerPipeline, DockerPipelineProps } from './docker-pipeline';